added web side API for create new user | cleaned up types to dedicated folder
