+++ Web/src/api/types/users.ts
@@ -1,0 +1,8 @@
+export interface CreateUserData {
+  email: string;
+  password: string;
+  role: "admin" | "doctor" | "patient" | "secretary";
+  name: string;
+  last_name: string;
+  national_id_number: string;
+}
